Created base class for all the Planning examples.
All that really differs in the example air cargo and cake baking examples are
the creation of the actions. So pass an action creation function into the class
constructor, and all the rest of the code can be shared by the other problems!

Implemented the Spare Tyre problem to give an example of how the new class and
action creation functions could be used; will convert other problems later.